# app/Http/Requests/Orders/UpdateOrderRequest.php
<?php

namespace App\Http\Requests\Orders;

use Illuminate\Foundation\Http\FormRequest;

class UpdateOrderRequest extends FormRequest
{
    /**
     * Determine if the user is authorized to make this request.
     */
    public function authorize(): bool
    {
        return true;
    }

    /**
     * Get the validation rules that apply to the request.
     *
     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
     */
    public function rules(): array
    {
        return [
            'customer_name' => 'min:3|max:32',
            'delivery_address',
            'order_total' => 'numeric|regex:/^\d{1,6}(\.\d{1,2})?$/',
            'order_status' => 'integer|in:0,1,2',
        ];
    }
}